+++ docs/changelog.md
@@ -5,6 +5,12 @@
 ## Changelog
 
 
+### v0.5.0 Alpha release
+
+- Replace scrollbars by mouse wheel operation.
+
+Based on PyWriter v9.0.5
+
 ### v0.4.0 Alpha release
 
 - Add mouse wheel support.